svn472
vs: zaple openmp

+++ src/RRCollider/RRCollider.cpp
@@ -5,6 +5,7 @@
 
 #include <assert.h>
 #include <memory.h>
+#include <omp.h> // error in msvc manifest code, needs omp.h even when using only pragmas
 #include <windows.h> // jen kvuli GetTempPath
 
 
@@ -77,7 +78,7 @@ RRCollider* RRCollider::create(RRMesh* importer, IntersectTechnique intersectTec
 void RRCollider::intersectBatch(RRRay* ray, unsigned numRays)
 {
 	#pragma omp parallel for schedule(static,1)
-	for(unsigned i=0;i<numRays;i++)
+	for(int i=0;i<(int)numRays;i++)
 	{
 		if(!intersect(ray+i)) ray[i].hitDistance = -1;
 	}
